feat: add --version flag to print out plugin version
this is a useful diagnostic feature; while implementing it
I discovered that the plugin main.go is already configured
to have goreleaser inject a version and commit property
via LDFLAGS.

I originally build the version_gen.go script to help expose
app version information as a package which could be
referenced in the code, rather than via LDFLAGS at the
top level but here the pattern seems to be to pass the
top-level version into the app constructor as a parameter
which is another way to do it.

I will remove in the next commit the version_gen.go pattern and
the version number from the build to try the terraform
plugin goreleaser pattern with a full heart; that way
it's more likely to align with community plugin developer
standards and expectations.